// front-endOficial/types/pacientesType.ts
export interface PacienteResumoType {
    id: number;
    nome: string;
    cpf: string;
    dataNascimento: string;
    criadoEm: string;
    prontuarioNumero: string;
    status: boolean;

}